most actions didn't over 255
Former-commit-id: b56e1444d5d533261a44f306dae8a76dc248ace5
Former-commit-id: ad8e05e21f78da753c39ad91fa7020b54d39a294

+++ JAxtar/astar.py
@@ -99,7 +99,7 @@ def astar_builder(
 
             neighbours, ncost = neighbours_fn(states, filled)
             parent_action = jnp.tile(
-                jnp.arange(ncost.shape[0], dtype=jnp.uint32)[:, jnp.newaxis], (1, ncost.shape[1])
+                jnp.arange(ncost.shape[0], dtype=jnp.uint8)[:, jnp.newaxis], (1, ncost.shape[1])
             )
             nextcosts = cost_val[jnp.newaxis, :] + ncost  # [n_neighbours, batch_size]
             filleds = jnp.isfinite(nextcosts)  # [n_neighbours, batch_size]
